% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_stats.R
\name{cor_diff}
\alias{cor_diff}
\title{Correlation Difference}
\usage{
cor_diff(cor1 = NULL, cor2 = NULL, n1 = NULL, n2 = NULL, x1 = NULL, x2 = NULL)
}
\arguments{
\item{cor1}{Numeric, the correlation coefficient for group 1.}

\item{cor2}{Numeric, the correlation coefficient for group 2.}

\item{n1}{Integer, the sample size for group 1.}

\item{n2}{Integer, the sample size for group 2.}

\item{x1}{A numeric vector for group 1.}

\item{x2}{A numeric vector for group 2.}
}
\value{
A data frame with the difference in correlations and the sampling variances of the difference.
}
\description{
Computes the difference in correlation coefficients between two groups along with associated sampling variance for meta-analysis. The correlation and associated sample size or the data frame of the two variables can be provided
}
\examples{
\dontrun{
set.seed(982)
# Example with known correlations
cor_diff(0.5, 0.3, 100, 100)
cor_diff(0.2, 0.4, 40, 40)
}
}
