% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat.eklof2012}
\alias{dat.eklof2012}
\title{eklof}
\format{A data frame :
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{}
}
\description{
Dataset from Eklof 2012 meta-analysis.
}
