% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchard_leave1out.R
\name{.add_ci_lines}
\alias{.add_ci_lines}
\title{Add Confidence Interval Lines to Orchard Plot}
\usage{
.add_ci_lines(orig_results, ci_lines_color)
}
\arguments{
\item{orig_results}{A data frame containing the original model's estimates. It must
include the columns \code{lowerCL} and \code{upperCL} representing the lower and upper
confidence limits, respectively.}

\item{ci_lines_color}{Character string specifying the color for the confidence interval lines.}
}
\value{
A \pkg{ggplot2} layer (via \code{geom_hline}) that can be added to a plot.
}
\description{
Internal function that creates a \pkg{ggplot2} layer to add horizontal lines
corresponding to the original model's 95% confidence limits.
}
\details{
These lines help visualize how the leave-one-out effect size estimates compare to the
overall model estimate.
}
\author{
Facundo Decunta - fdecunta@agro.uba.ar
}
\keyword{internal}
