% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchard_plot.R
\name{.add_colour_blind_palette}
\alias{.add_colour_blind_palette}
\title{Set colour blind friendly palette}
\usage{
.add_colour_blind_palette(plt)
}
\description{
Set colour blind friendly palette
}
\keyword{internal}
