% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchard_plot.R
\name{.add_conf_intervals}
\alias{.add_conf_intervals}
\title{Add confidence intervals}
\usage{
.add_conf_intervals(plt, mod_table, lines_position, branch.size)
}
\description{
Add confidence intervals
}
\keyword{internal}
