% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchard_leave1out.R
\name{.add_ghost_points}
\alias{.add_ghost_points}
\title{Add Ghost Points to Orchard Plot}
\usage{
.add_ghost_points(data, transfm)
}
\arguments{
\item{data}{Data from the leave-one-out analysis, specifically the \code{data} component
containing effect sizes (yi), variances (vi), and grouping variable (stdy).}

\item{transfm}{Character string indicating the transformation to be applied to the data.
See \code{orchard_leave1out()} for available transformation options.}
}
\value{
A \pkg{ggplot2} layer (via \code{geom_quasirandom}) displaying the ghost points.
}
\description{
Internal function to add ghost points to an orchard plot.
}
\details{
Ghost points represent the effect sizes that were omitted in each leave-one-out iteration.
They are plotted as empty points with reduced opacity to indicate their auxiliary role.
}
\author{
Facundo Decunta - fdecunta@agro.uba.ar
}
\keyword{internal}
