% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchard_plot.R
\name{.add_point_estimates}
\alias{.add_point_estimates}
\title{Add point estimates for orchard plot}
\usage{
.add_point_estimates(
  plt,
  mod_table,
  color2,
  points_shapes,
  lines_position,
  trunk.size,
  shapes_values
)
}
\description{
Add point estimates for orchard plot
}
\keyword{internal}
