% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchard_plot.R
\name{.add_pred_intervals}
\alias{.add_pred_intervals}
\title{Add prediction intervals}
\usage{
.add_pred_intervals(plt, mod_table, lines_position, trunk.size, twig.size)
}
\description{
Add prediction intervals
}
\keyword{internal}
