% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchard_plot.R
\name{.add_reference_line}
\alias{.add_reference_line}
\title{Add reference line for orchard plot}
\usage{
.add_reference_line(plt, alpha)
}
\description{
Add reference line for orchard plot
}
\keyword{internal}
