% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{.apply_transformation}
\alias{.apply_transformation}
\title{Apply transformation to results}
\usage{
.apply_transformation(results, n_transfm, transfm)
}
\description{
Apply transformation to results
}
\keyword{internal}
