% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{.base_bubble_plot}
\alias{.base_bubble_plot}
\title{Create the base bubble plot layer.}
\usage{
.base_bubble_plot(data_trim, alpha)
}
\description{
Create the base bubble plot layer.
}
\keyword{internal}
