% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchard_plot.R
\name{.base_orchard_plot}
\alias{.base_orchard_plot}
\title{Create base for orchard plot}
\usage{
.base_orchard_plot(data_trim, colour, fill, alpha)
}
\description{
Create base for orchard plot
}
\keyword{internal}
