% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{.bbp_conf_interval}
\alias{.bbp_conf_interval}
\title{Add confidence interval lines to bubble plot.}
\usage{
.bbp_conf_interval(mod_table, ci.lwd, ci.col)
}
\description{
Add confidence interval lines to bubble plot.
}
\keyword{internal}
