% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{.bbp_estimate_line}
\alias{.bbp_estimate_line}
\title{Add estimate line to bubble plot.}
\usage{
.bbp_estimate_line(mod_table, est.lwd, est.col)
}
\description{
Add estimate line to bubble plot.
}
\keyword{internal}
