% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{.bbp_facets}
\alias{.bbp_facets}
\title{Add facetting to bubble plot by condition.}
\usage{
.bbp_facets(data_trim, condition.nrow, condition.order)
}
\description{
Add facetting to bubble plot by condition.
}
\keyword{internal}
