% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{.bbp_kg_labels}
\alias{.bbp_kg_labels}
\title{Add k and g annotation labels to bubble plot.}
\usage{
.bbp_kg_labels(k, g, k.pos, kg_labels)
}
\description{
Add k and g annotation labels to bubble plot.
}
\keyword{internal}
