% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{.bbp_legends}
\alias{.bbp_legends}
\title{Add and position legends on bubble plot.}
\usage{
.bbp_legends(legend.pos, size_legend)
}
\description{
Add and position legends on bubble plot.
}
\keyword{internal}
