% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{.bbp_pred_interval}
\alias{.bbp_pred_interval}
\title{Add prediction interval lines to bubble plot.}
\usage{
.bbp_pred_interval(mod_table, pi.lwd, pi.col)
}
\description{
Add prediction interval lines to bubble plot.
}
\keyword{internal}
