% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_stats.R
\name{.calc.kurtosis}
\alias{.calc.kurtosis}
\title{.calc.kurtosis}
\usage{
.calc.kurtosis(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric value representing the kurtosis estimate or variance.
}
\description{
Computes the excess kurtosis of a numeric vector using small sample correction.
}
