% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_stats.R
\name{.calc.skewness}
\alias{.calc.skewness}
\title{.calc.skewness}
\usage{
.calc.skewness(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric value representing the skewness estimate or variance.
}
\description{
Computes the skewness of a numeric vector using small sample correction.
}
