% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\docType{data}
\name{.colour_blind_palette}
\alias{.colour_blind_palette}
\title{Colour blind palette}
\format{
An object of class \code{character} of length 20.
}
\usage{
.colour_blind_palette
}
\description{
Colour blind palette
}
\keyword{intern}
