% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.column_exists}
\alias{.column_exists}
\title{Check if Column Exists}
\usage{
.column_exists(data, column_name)
}
\arguments{
\item{data}{data.frame. A data frame to check}

\item{column_name}{character. The name of the column to look for}
}
\value{
logical. TRUE if the column exists, FALSE otherwise
}
\description{
Verifies if a specified column exists in the data frame
}
\author{
Facundo Decunta - fdecunta@agro.uba.ar
}
\keyword{internal}
