% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.column_is_categorical}
\alias{.column_is_categorical}
\title{Check if Column is Categorical}
\usage{
.column_is_categorical(col)
}
\arguments{
\item{col}{vector. A vector to check (character, factor, or numeric)}
}
\value{
logical. TRUE if the column can be used as a category, FALSE otherwise
}
\description{
Determines if a column can be used as a categorical variable.
Valid categorical variables include:
- Character vectors
- Factors
- Integer values (common for ID numbers used as factors)
}
\author{
Facundo Decunta - fdecunta@agro.uba.ar
}
\keyword{internal}
