% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leave_one_out.R
\name{.create_tmp_vcv}
\alias{.create_tmp_vcv}
\title{Create Temporary Variance-Covariance Matrix}
\usage{
.create_tmp_vcv(data, vcalc_args)
}
\arguments{
\item{data}{A data frame containing the variables specified in vcalc_args.}

\item{vcalc_args}{A list of arguments for metafor::vcalc function, including:
\itemize{
  \item vi: Name of the variance column
  \item cluster: Name of the clustering variable column
  \item obs: Name of the observation ID column
  \item rho: Correlation coefficient between effect sizes
}}
}
\value{
A variance-covariance matrix for use in meta-analytic models.
}
\description{
Creates a variance-covariance matrix for a subset of data using metafor's vcalc function.
}
\author{
Facundo Decunta - fdecunta@agro.uba.ar
}
\keyword{internal}
