% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{.extract_model_data}
\alias{.extract_model_data}
\title{Extract model data used in metafor model}
\usage{
.extract_model_data(model)
}
\arguments{
\item{model}{A \code{metafor} model object.}
}
\value{
A data frame of the used observations.
}
\description{
Gets the data actually used to fit a \pkg{metafor} model, after applying
any subset and removing rows with missing values.
}
\keyword{internal}
