% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leave_one_out.R
\name{.get_effectsizes}
\alias{.get_effectsizes}
\title{Get Leave-One-Out Effect Sizes}
\usage{
.get_effectsizes(outputs, group)
}
\arguments{
\item{outputs}{A named list of model objects from leave-one-out analysis.}

\item{group}{A string specifying the grouping variable used in the analysis.}
}
\value{
A data frame of effect sizes with a column indicating the omitted group.
}
\description{
Extracts and aggregates effect size data from each leave-one-out iteration.
}
\author{
Facundo Decunta - fdecunta@agro.uba.ar
}
\keyword{internal}
