% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R, R/plot_helpers.R
\name{.get_kg_labels}
\alias{.get_kg_labels}
\title{Compute k and g labels for each condition.}
\usage{
.get_kg_labels(dat)

.get_kg_labels(dat)
}
\description{
Takes data from mod_results and compute the K (number of effect sizes)
and G (number of groups). 
NOTE: It expects the condition from data to be a factor
}
\keyword{internal}
