% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{.get_results}
\alias{.get_results}
\title{Get the output from mod_results}
\usage{
.get_results(object, mod, group, N, by, at, weights)
}
\description{
Take the inputs passed to plotting functions and make sure to convert them
into the results from \code{mod_results}.
}
\keyword{internal}
