% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_metafor_object}
\alias{.is_metafor_object}
\title{Check if Object is from metafor Package}
\usage{
.is_metafor_object(obj)
}
\arguments{
\item{obj}{object. An R object to check for metafor class inheritance}
}
\value{
logical. TRUE if the object is from metafor package, FALSE otherwise
}
\description{
Verifies if an object is of class rma.mv, rma, rma.uni, or robust.rma
}
\keyword{internal}
