% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_stats.R
\name{.lnM}
\alias{.lnM}
\title{.lnM}
\usage{
.lnM(x1, x2, sd1, sd2, n1, n2)
}
\arguments{
\item{x1}{Mean of group 1.}

\item{x2}{Mean of group 2.}

\item{sd1}{Standard deviation of group 1.}

\item{sd2}{Standard deviation of group 2.}

\item{n1}{Sample size of group 1.}

\item{n2}{Sample size of group 2.}
}
\value{
The natural logarithm of the ratio of between and within group variances.
}
\description{
Computes the natural logarithm of the ratio of between and within group variances as a measure of the magnitude of difference.
}
