% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_stats.R
\name{.MSb}
\alias{.MSb}
\title{.MSb}
\usage{
.MSb(x1, x2, n1, n2, paired = FALSE)
}
\arguments{
\item{x1}{Mean of group 1.}

\item{x2}{Mean of group 2.}

\item{n1}{Sample size of group 1.}

\item{n2}{Sample size of group 2.}

\item{paired}{Logical, whether the samples are paired. Default is FALSE.}
}
\value{
The between group mean-squares difference.
}
\description{
Computes the between group mean-squares difference.
}
