% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_stats.R
\name{.MSw}
\alias{.MSw}
\title{.MSw}
\usage{
.MSw(sd1, sd2, n1, n2, paired = FALSE)
}
\arguments{
\item{sd1}{Standard deviation of group 1.}

\item{sd2}{Standard deviation of group 2.}

\item{n1}{Sample size of group 1.}

\item{n2}{Sample size of group 2.}

\item{paired}{Logical, whether the samples are paired. Default is FALSE.}
}
\value{
The within group mean-squares difference.
}
\description{
Computes the within group mean-squares difference.
}
