% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchard_plot.R
\name{.orcd_colour_blind_palette}
\alias{.orcd_colour_blind_palette}
\title{Set a colour-blind-friendly palette}
\usage{
.orcd_colour_blind_palette(cb)
}
\description{
Set a colour-blind-friendly palette
}
\keyword{internal}
