% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchard_plot.R
\name{.orcd_conf_intervals}
\alias{.orcd_conf_intervals}
\title{Add confidence intervals}
\usage{
.orcd_conf_intervals(mod_table, lines_position, branch.size)
}
\description{
Add confidence intervals
}
\keyword{internal}
