% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchard_plot.R
\name{.orcd_point_estimates}
\alias{.orcd_point_estimates}
\title{Add point estimates for orchard plot}
\usage{
.orcd_point_estimates(mod_table, colour, trunk.size)
}
\description{
Note: Point estimate previously used 'geom_pointrange',
plotting the estimate and prediction interval at the same time.
However, that was less flexible.
Now PI are plotted with 'geom_linerange' and point estimate with 'geom_point'.
}
\keyword{internal}
