% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchard_plot.R
\name{.orcd_pred_intervals}
\alias{.orcd_pred_intervals}
\title{Add prediction intervals}
\usage{
.orcd_pred_intervals(mod_table, lines_position, trunk.size, twig.size)
}
\description{
Add prediction intervals
}
\keyword{internal}
