% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchard_plot.R
\name{.orcd_reference_line}
\alias{.orcd_reference_line}
\title{Add reference line for orchard plot}
\usage{
.orcd_reference_line(alpha)
}
\description{
Add reference line for orchard plot
}
\keyword{internal}
