% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchard_plot.R
\name{.order_mod}
\alias{.order_mod}
\title{Reorder the tree}
\usage{
.order_mod(results, mod.order)
}
\description{
Reorder the position of the moderators in the plot by reordering
the factors in the model table and the data.
}
\keyword{internal}
