% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leave_one_out.R
\name{.prune_tree}
\alias{.prune_tree}
\title{Prune Tree For Leave-One-Out}
\usage{
.prune_tree(tree, species_names)
}
\arguments{
\item{tree}{A phylogenetic tree object.}

\item{species_names}{A vector of species names.}
}
\description{
Prune Tree For Leave-One-Out
}
\author{
Daniel Noble  - daniel.noble@anu.edu.au

Facundo Decunta - fdecunta@agro.uba.ar
}
