% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_stats.R
\name{.r.to.zr}
\alias{.r.to.zr}
\title{Fisher Z-transformation of Correlation Coefficient}
\usage{
.r.to.zr(r)
}
\arguments{
\item{r}{A numeric vector of Pearson correlation coefficients.}
}
\value{
The Zr-transformed correlation value(s).
}
\description{
Converts a Pearson correlation coefficient \eqn{r} to Fisher's Zr using the inverse hyperbolic tangent.
}
