% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{.set_condition}
\alias{.set_condition}
\title{Set the condition variable for plotting.}
\usage{
.set_condition(results, condition.levels)
}
\description{
Set the condition variable for plotting.
}
\keyword{internal}
