% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_stats.R
\name{.sv_jack}
\alias{.sv_jack}
\title{Jackknife Estimation of Skewness and Kurtosis for computing sampling variances}
\usage{
.sv_jack(
  x,
  bias.correct = TRUE,
  return.replicates = FALSE,
  type = c("skew", "kurt")
)
}
\arguments{
\item{x}{A numeric vector.}

\item{bias.correct}{Logical, whether to apply bias correction. Default is TRUE.}

\item{return.replicates}{Logical, whether to return jackknife replicates. Default is FALSE.}

\item{type}{Character, either "skew" or "kurt" to specify the type of statistic.}
}
\value{
A list with estimate (`est`), bias-corrected estimate (`est_bc`), variance, and standard error.
}
\description{
Computes jackknife estimate and standard error for skewness and kurtosis.
}
