% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_stats.R
\name{.v_lnM}
\alias{.v_lnM}
\title{.v_lnM}
\usage{
.v_lnM(x1, x2, sd1, sd2, n1, n2, r = NULL)
}
\arguments{
\item{x1}{Mean of group 1.}

\item{x2}{Mean of group 2.}

\item{sd1}{Standard deviation of group 1.}

\item{sd2}{Standard deviation of group 2.}

\item{n1}{Sample size of group 1.}

\item{n2}{Sample size of group 2.}

\item{r}{Optional correlation between groups for dependent samples. Default is NULL for independent samples.}
}
\value{
The sampling variance of the natural logarithm of the ratio of between and within group variances.
}
\description{
Computes the sampling variance of the natural logarithm of the ratio of between and within group variances.
}
