% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leave_one_out.R
\name{.validate_phylo_args}
\alias{.validate_phylo_args}
\title{Validate Phylogenetic Arguments}
\usage{
.validate_phylo_args(model, phylo_args)
}
\arguments{
\item{model}{A metafor model object with random effects.}

\item{phylo_args}{A list containing the arguments for the phylogenetic matrix calculation:
\itemize{
  \item tree: A phylogenetic tree object of class "phylo"
  \item species_colname: Name of the column in the data corresponding to species names,
        which should be a random factor linked to a matrix in the model
}}
}
\value{
The validated phylo_args list if all checks pass.
}
\description{
Checks the validity of the arguments provided for phylogenetic matrix calculations.
Ensures the tree is a proper phylogenetic object and that the species column name
is correctly linked to a random effect in the model.
}
\keyword{internal}
