% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leave_one_out.R
\name{.validate_robust_args}
\alias{.validate_robust_args}
\title{Validate robust_args}
\usage{
.validate_robust_args(model_data, robust_args)
}
\arguments{
\item{model_data}{A data frame containing the variables used in the model.}

\item{robust_args}{A list of arguments for the metafor::robust function.}
}
\value{
The validated robust_args list.
}
\description{
Validates that robust_args contains the required parameters and that they
reference valid columns in the data.
}
\author{
Facundo Decunta - fdecunta@agro.uba.ar
}
\keyword{internal}
