% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_stats.R
\name{.zr.variance}
\alias{.zr.variance}
\title{Variance of Fisher Z-transformed Correlation}
\usage{
.zr.variance(n)
}
\arguments{
\item{n}{Sample size (must be greater than 3).}
}
\value{
The variance of the Zr-transformed correlation.
}
\description{
Computes the approximate sampling variance of Fisher's Zr given a sample size.
}
