% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eklof}
\alias{eklof}
\title{eklof}
\format{
A data frame :
\describe{
  \item{ExptID}{Experiment / effect size ID}
  \item{First.author}{First author of publication}
	 \item{Publication.year}{Year of publication}
	 \item{Journal}{Journal published within}
	 \item{Research.group}{Research group that data arose from}
	 \item{Grazer.type}{Type pf grazer}
	 \item{mean_treatment}{Mean abundance/biomass of invertebrates in treatment group}
	 \item{SD_treatment}{Standard deviation in abundance/biomass of invertebrates in treatment group}
	 \item{N_treatment}{Sample size of invertebrate sample in treatment group}
	 \item{mean_control}{Mean abundance/biomass of invertebrates in control group}
	 \item{SD_control}{Standard deviation in abundance/biomass of invertebrates in control group}
	 \item{N_control}{Sample size of invertebrate sample in control group}
  ...
}
}
\description{
Eklof et al. (2012) evaluated the effects of predation on benthic invertebrate communities. Using the log response ratio they quantified differences in abundance and/or biomass of gastropods and Amphipods in groups with and without predation in an experimental setting. Below we describe the variables used in our examples and analyses.
}
\references{
Eklof J.S., Alsterberg C., Havenhand J.N., Sundback K., Wood H.L., Gamfeldt L. 2012. Experimental climate change weakens the insurance effect of biodiversity. Ecology Letters, 15:864-872. https://doi.org/10.1111/j.1461-0248.2012.01810.x
}
