% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{firstup}
\alias{firstup}
\title{firstup}
\usage{
firstup(x, upper = TRUE)
}
\arguments{
\item{x}{a character string}

\item{upper}{logical indicating if the first letter of the character string should be capitalized. Defaults to TRUE.}
}
\value{
Returns a character string with all combinations of the moderator level names with upper case first letters
}
\description{
Uppercase moderator names
}
\author{
Shinichi Nakagawa - s.nakagawa@unsw.edu.au

Daniel Noble - daniel.noble@anu.edu.au
}
