% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data(model, mod)
}
\arguments{
\item{model}{rma.mv object}

\item{mod}{the moderator variable}
}
\value{
Returns a data frame
}
\description{
Collects and builds the data used to fit the rma.mv or rma model in metafor
}
