% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{get_data2}
\alias{get_data2}
\title{get_data2}
\usage{
get_data2(model, mod, data = data)
}
\arguments{
\item{model}{rma.mv object}

\item{mod}{the moderator variable}
}
\value{
Returns a data frame
}
\description{
Collects and builds the data used to fit the rma.mv or rma model in metafor in conjunction with emmeans
}
\author{
Shinichi Nakagawa - s.nakagawa@unsw.edu.au

Daniel Noble - daniel.noble@anu.edu.au
}
