% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{get_pred}
\alias{get_pred}
\title{get_pred}
\usage{
get_pred(model, mod)
}
\arguments{
\item{model}{rma.mv object}

\item{mod}{the name of a moderator}
}
\description{
Function to get prediction intervals (crediblity intervals) from rma objects (metafor)
}
\author{
Shinichi Nakagawa - s.nakagawa@unsw.edu.au

Daniel Noble - daniel.noble@anu.edu.au
}
