% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{is_categorical}
\alias{is_categorical}
\title{Check if an Object is Categorical}
\usage{
is_categorical(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
A logical value: `TRUE` if `x` is a character vector, a factor, or `NULL`; otherwise, `FALSE`.
}
\description{
Determines whether the given object is categorical, defined as a character
vector, a factor, or `NULL`.
}
\keyword{internal}
