% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leave_one_out.R
\name{leave_one_out}
\alias{leave_one_out}
\title{Leave-One-Out Analysis for Meta-Analytic Models}
\usage{
leave_one_out(
  model,
  group,
  vcalc_args = NULL,
  robust_args = NULL,
  phylo_args = NULL
)
}
\arguments{
\item{model}{A meta-analytic model fitted using **metafor**.}

\item{group}{A factor or categorical variable specifying the leave-one-out groups.}

\item{vcalc_args}{Optional list of arguments for the variance-covariance calculation using 
metafor's vcalc function. Must include 'vi', 'cluster', 'obs', and 'rho' elements.}

\item{robust_args}{Optional list of arguments for robust variance estimation using
metafor's robust function. Must include a 'cluster' element.}
}
\value{
Same as `mod_results()`, but with the estimates from each model ran
  in the leave-one-out analysis, and the effect sizes from each model.
}
\description{
Performs a leave-one-out analysis on a meta-analytic model from
the **metafor** package by iteratively removing each level of a grouping
variable and refitting the model.
}
\examples{
\dontrun{
res <- metafor::rma.mv(lnrr, lnrr_vi, random = ~ 1 | paper_ID, data = fish)
loo_results <- leave_one_out(res, group = "paper_ID")

# With variance-covariance calculation
loo_results <- leave_one_out(res, group = "paper_ID", 
                             vcalc_args = list(vi = "lnrr_vi", 
                                              cluster = "paper_ID", 
                                              obs = "es_ID", 
                                              rho = 0.5))
                                              
# With robust variance estimation
loo_results <- leave_one_out(res, group = "paper_ID", 
                             robust_args = list(cluster = "paper_ID"))
}

}
\author{
Facundo Decunta - fdecunta@agro.uba.ar
}
