% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m1_ml.R
\name{m1_ml}
\alias{m1_ml}
\title{m1_ml}
\usage{
m1_ml(model, boot = NULL)
}
\arguments{
\item{model}{Model object of class \code{rma.mv} or \code{rma}. Currently only model objects using the \code{mods} argument work (e.g., \code{mod = ~1}).}

\item{boot}{Number of simulations to run to produce 95 percent confidence intervals for M1. Default is \code{NULL}, where only the point estimate is provided.}
}
\value{
A data frame containing all the model results including mean effect size estimate, confidence, and prediction intervals
}
\description{
M1 for mulilevel meta-analytic models, based on Yang et al. (2023). Under multilevel models, we can have multiple M1 - TODO - we need to cite original M1 paper
}
\examples{
\donttest{
library(metafor)
# NOTE: boot is set LOW here for speed; use boot >= 1000 in practice.
data(english)
english <- escalc(measure = "SMD", n1i = NStartControl,
  sd1i = SD_C, m1i = MeanC, n2i = NStartExpt, sd2i = SD_E,
  m2i = MeanE, var.names = c("SMD", "vSMD"), data = english)
english_MA <- rma.mv(yi = SMD, V = vSMD,
  random = list(~1 | StudyNo, ~1 | EffectID), data = english)
m1_ml(english_MA)
m1_ml(english_MA, boot = 10)
}
}
\references{
TODO
}
\author{
Shinichi Nakagawa - s.nakagawa@unsw.edu.au

Daniel Noble - daniel.noble@anu.edu.au
}
