% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{marginal_means}
\alias{marginal_means}
\title{marginal_means}
\usage{
marginal_means(
  model,
  data,
  mod = "1",
  weights = "prop",
  by = NULL,
  at = NULL,
  ...
)
}
\arguments{
\item{model}{rma.mv object}

\item{data}{data frame used to fit rma.mv model}

\item{mod}{moderator variable of interest that one wants marginal means for.}

\item{weights}{how to marginalize categorical variables. The default is weights = "prop" - more to add here (discuss with Dan)}
}
\description{
Function to to get marginal means from met-regression models with single or multiple moderator variables that are both continuous or categorical.
}
\author{
Shinichi Nakagawa - s.nakagawa@unsw.edu.au

Daniel Noble - daniel.noble@anu.edu.au
}
