% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{marginalised_means}
\alias{marginalised_means}
\title{marginalised_means}
\usage{
marginalised_means(model, dat, pred, by = NULL, at = NULL, ...)
}
\arguments{
\item{model}{rma.mv object}

\item{dat}{data frame used to fit rma.mv model}

\item{pred}{predictor variable of interest that one wants marginalised means for.}
}
\description{
Function to to get marginalised means from met-regression models with single or multiple moderator variables that are both continuous or categorical.
}
\author{
Shinichi Nakagawa - s.nakagawa@unsw.edu.au

Daniel Noble - daniel.noble@anu.edu.au
}
