% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i2_ml.R
\name{matrix_i2}
\alias{matrix_i2}
\title{matrix_i2}
\usage{
matrix_i2(model)
}
\arguments{
\item{model}{Model object of class \code{rma.mv} or \code{rma}.}
}
\value{
A named numeric vector of I2 (I-squared) values expressed as percentages. The first element, \code{I2_Total}, is the total heterogeneity; each remaining element (named \code{I2_<level>} after a random-effect level of the model) is the heterogeneity attributable to that level.
}
\description{
Calculated I2 (I-squared) for mulilevel meta-analytic models, based on a matrix method proposed by Wolfgang Viechtbauer.
}
\examples{
\donttest{
library(metafor)
data(english)
english <- escalc(
  measure = "SMD", n1i = NStartControl,
  sd1i = SD_C, m1i = MeanC,
  n2i = NStartExpt, sd2i = SD_E, m2i = MeanE,
  var.names = c("SMD", "vSMD"), data = english)
english_MA <- rma.mv(
  yi = SMD, V = vSMD,
  random = list(~1 | StudyNo, ~1 | EffectID),
  data = english)
matrix_i2(english_MA)
}
}
