% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{mod_is_valid}
\alias{mod_is_valid}
\title{Validate 'mod'}
\usage{
mod_is_valid(model, mod)
}
\arguments{
\item{model}{A meta-analytic model from the \pkg{metafor} package.}

\item{mod}{A string specifying the moderator to check. Must be `"1"` or one of 
the moderators included in \code{model$formula.mods}.}
}
\value{
A logical value: \code{TRUE} if `mod` is valid, strop and throw an error otherwise.
}
\description{
Checks if `mod` argument is valid. This argument must be either
`"1"` (indicating an intercept-only model) or a moderator that is included in the
meta-analytic model (as specified in `model$formula.mods`).
}
\keyword{internal}
