% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{mod_results}
\alias{mod_results}
\title{mod_results}
\usage{
mod_results(model, mod, group, data)
}
\arguments{
\item{model}{rma.mv or rma object}

\item{mod}{the name of a moderator; put "1" if the intercept model (meta-analysis) or no moderators.}

\item{group}{The grouping variable that one wishes to plot beside total effect sizes, k. This could be study, species or whatever other grouping variable one wishes to present sample sizes.}

\item{data}{The data frame used to fit the rma.mv model object}
}
\value{
A data frame containing all the model results including mean effect size estimate, confidence and prediction intervals
}
\description{
Using a metafor model object of class rma or rma.mv it creates a table of model results containing the mean effect size estimates for all levels of a given categorical moderator, their corresponding confidence intervals and prediction intervals
}
\examples{
\dontrun{data(eklof)
eklof<-metafor::escalc(measure="ROM", n1i=N_control, sd1i=SD_control,
m1i=mean_control, n2i=N_treatment, sd2i=SD_treatment, m2i=mean_treatment,
data=eklof)
# Add the unit level predictor
eklof$Datapoint<-as.factor(seq(1, dim(eklof)[1], 1))
# fit a MLMR - accouting for some non-independence
eklof_MR<-metafor::rma.mv(yi=yi, V=vi, mods=~ Grazer.type, random=list(~1|ExptID,
~1|Datapoint), data=eklof)
results <- mod_results(eklof_MR, mod = "Grazer.type", group = "ExptID", data=eklof)
}
}
\author{
Shinichi Nakagawa - s.nakagawa@unsw.edu.au

Daniel Noble - daniel.noble@anu.edu.au
}
