% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{model_is_valid}
\alias{model_is_valid}
\title{Validate 'model'}
\usage{
model_is_valid(model)
}
\arguments{
\item{model}{An object representing a fitted metafor model. It should be of class
`robust.rma`, `rma.mv`, `rma`, or `rma.uni`.}
}
\value{
Logical `TRUE` if the `model` argument is valid. If the model is missing or
  not of an accepted class, the function stops with an error message.
}
\description{
Checks if `model` is a valid metafor model object.
The model must be one of the following classes: `robust.rma`, `rma.mv`, `rma`, or `rma.uni`.
}
\keyword{internal}
