% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leave_one_out.R
\name{models_to_dataframe}
\alias{models_to_dataframe}
\title{Convert List Of Model Outputs To Data Frame}
\usage{
models_to_dataframe(models_outputs)
}
\arguments{
\item{models_outputs}{A named list of models, where each model is expected to 
have elements `b`, `ci.lb`, and `ci.ub` representing the coefficient and
its confidence interval.}
}
\value{
A data frame with the following columns:
  - `left_out`: Names of the models (from `models_outputs`).
  - `b`: The extracted coefficient.
  - `ci_lb`: The lower bound of the confidence interval.
  - `ci_ub`: The upper bound of the confidence interval.
}
\description{
Extracts the first coefficient and its confidence interval
from a list of model outputs and converts them into a structured data frame.
}
\details{
The function assumes that each model in `models_outputs` has
  a structure with elements `b`, `ci.lb`, and `ci.ub`. The function extracts
  the first coefficient and its corresponding confidence interval and
  returns them in a tidy format.
}
\keyword{internal}
