% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_stats.R
\name{moment_effects}
\alias{moment_effects}
\title{moment_effects}
\usage{
moment_effects(x1, x2, type = c("skew", "kurt"))
}
\arguments{
\item{x1}{A numeric vector.}

\item{x2}{A numeric vector.}

\item{type}{Character, either "skew" or "kurt" to specify the type of moment effect.}
}
\value{
A data frame with the difference in moment effects and their variances.
}
\description{
Computes the effect estimate of the difference in skewness and kurtosis between two groups along with associated sampling variance for meta-analysis.
}
\examples{
\dontrun{
set.seed(982)
library(moments, PearsonDS)

# Just some random comparisons
moment_effects(rnorm(100), rnorm(100), type = "skew")
moment_effects(rnorm(40), rnorm(40), type = "skew")
moment_effects(rnorm(40), rnorm(40), type = "kurt")
moment_effects(rnorm(100), rnorm(100), type = "kurt")

# Comparisons with known moment differences
m1 <- c(mean = 0, variance = 1, skewness = 0, kurtosis = 2)
m2 <- c(mean = 0, variance = 1, skewness = 0, kurtosis = 4)

x1 <- PearsonDS::rpearson(1000, moments = m1)
x2 <- PearsonDS::rpearson(1000, moments = m2)
moment_effects(x1, x2, type = "skew") # ~0
moment_effects(x1, x2, type = "kurt") # ~-2

m1 <- c(mean = 0, variance = 1, skewness = 0, kurtosis = 3)
m2 <- c(mean = 0, variance = 1, skewness = 1, kurtosis = 3)
x1 <- PearsonDS::rpearson(1000, moments = m1)
x2 <- PearsonDS::rpearson(1000, moments = m2)
moment_effects(x1, x2, type = "skew") # ~-1
moment_effects(x1, x2, type = "kurt") # ~0
}
}
