% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{num_studies}
\alias{num_studies}
\title{num_studies}
\usage{
num_studies(data, mod, group)
}
\arguments{
\item{data}{Raw data from object of class "orchard"}

\item{mod}{Character string describing the moderator of interest.}

\item{group}{A character string specifying the column name of the study ID grouping variable.}
}
\value{
Returns a table with the number of studies in each level of all parameters within a rma.mv or rma object.
}
\description{
Computes how many studies are in each level of categorical moderators of a rma.mv model object.
}
\examples{
\dontrun{data(fish)
warm_dat <- fish
model <- metafor::rma.mv(yi = lnrr, V = lnrr_vi, random = list( ~1 | es_ID,~1 | group_ID), mods = ~experimental_design-1, method = "REML", test = "t", data = warm_dat, control=list(optimizer="optim", optmethod="Nelder-Mead"))
num_studies(model$data, experimental_design, group_ID)
}
}
\author{
Shinichi Nakagawa - s.nakagawa@unsw.edu.au

Daniel Noble - daniel.noble@anu.edu.au
}
