% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phylo_matrix}
\alias{phylo_matrix}
\title{phylo_matrix}
\format{
A phylogenetic correlation matrix :
}
\description{
Phylogenetic matrix used in Pottier et al. (2021) meta-analysis on sex-differences in acclimation
}
\references{
Pottier et al. 2022. Functional Ecology
}
