% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{pred_interval_esmeans}
\alias{pred_interval_esmeans}
\title{pred_interval_esmeans}
\usage{
pred_interval_esmeans(model, mm, mod, ...)
}
\arguments{
\item{model}{rma.mv object}

\item{mm}{result from emmeans::emmeans object'}

\item{mod}{Moderator of interest}

\item{...}{other arguments passed to function}
}
\description{
Function to get prediction intervals (credibility intervals) from esmeans objects (metafor)
}
\author{
Shinichi Nakagawa - s.nakagawa@unsw.edu.au

Daniel Noble - daniel.noble@anu.edu.au
}
