% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{print.orchard}
\alias{print.orchard}
\title{print.orchard}
\usage{
\method{print}{orchard}(object, ...)
}
\arguments{
\item{object}{x an R object of class orchard}
}
\value{
Returns a data frame
}
\description{
Print method for class 'orchard'
}
\author{
Shinichi Nakagawa - s.nakagawa@unsw.edu.au

Daniel Noble - daniel.noble@anu.edu.au
}
