% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub_bias_plot.R
\name{pub_bias_plot}
\alias{pub_bias_plot}
\title{pub_bias_plot}
\usage{
pub_bias_plot(
  plot,
  fe_model,
  v_model = NULL,
  col = c("red", "blue"),
  plotadj = -0.05,
  textadj = 0.05,
  branch.size = 1.2,
  trunk.size = 3
)
}
\arguments{
\item{plot}{The orchard plot object to add the bias corrected mean and confidence intervals to. This plot needs to be a plot that displays the raw and a single meta-analytic mean (overall mean), confidence interval and predcition interval.}

\item{fe_model}{The meta-analytic model (rma object) produced from a two-step correction (sensu Yang et al. 2023) (Step 1: Fixed effect model) with a robust corretion to correct the meta-analytic mean for publication bias (selection bias) and the dependency in the data.}

\item{v_model}{An optional argument. The meta-analytic model (rma object) to deal with publication bias using the meta-regression approach proposed by Nakagawa et al. 2023. This model can have fixed and/or random effects, with one random effect being sampling error (se or v). The intercept from this model can be considred the corrected meta-analytic mean when sample size is infinite or sampling error is zero.}

\item{col}{The colour of the mean and confidence intervals.}

\item{plotadj}{The adjustment to the x-axis position of the mean and confidence intervals.}

\item{textadj}{The adjustment to the y-axis position of the mean and confidence intervals for the text displaying the type of correction.}

\item{branch.size}{Size of the confidence intervals.}

\item{trunk.size}{Size of the mean, or central point.}
}
\value{
An orchard plot with the corrected meta-analytic mean and confidence intervals added.
}
\description{
This function adds the bias corrected mean and confidence intervals to an existing orchard plot that displays the overall meta-analytic mean effect size.
}
