% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_ml.R
\name{R2_calc}
\alias{R2_calc}
\title{R2_calc}
\usage{
R2_calc(model)
}
\arguments{
\item{model}{Model object of class 'rma.mv', 'rma'}
}
\description{
Calculated R2 (R-squared) for mixed (mulitlevel) models, based on Nakagawa & Schielzeth (2013)
}
\examples{
\dontrun{
data(lim)
lim$vi<-(1/sqrt(lim$N - 3))^2
lim_MR<-metafor::rma.mv(yi=yi, V=vi, mods=~Phylum-1,
random=list(~1|Article, ~1|Datapoint), data=lim)
R2 <- R2_calc(lim_MR)
}
}
\author{
Shinichi Nakagawa - s.nakagawa@unsw.edu.au

Daniel Noble - daniel.noble@anu.edu.au
}
