% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_stats.R
\name{safe_lnM_dep}
\alias{safe_lnM_dep}
\title{SAFE bootstrap : dependent samples}
\usage{
safe_lnM_dep(
  x1bar,
  x2bar,
  sd1,
  sd2,
  n,
  r,
  min_kept = 2000,
  chunk_init = 4000,
  chunk_max = 2e+06,
  max_draws = Inf,
  patience_noaccept = 5
)
}
\arguments{
\item{x1bar}{Mean of group 1.}

\item{x2bar}{Mean of group 2.}

\item{sd1}{Standard deviation of group 1.}

\item{sd2}{Standard deviation of group 2.}

\item{n}{Sample size (assumed equal for both groups).}

\item{r}{Correlation between the two groups.}

\item{min_kept}{Minimum number of accepted bootstrap samples to keep. Default is 2000.}

\item{chunk_init}{Initial chunk size for bootstrap sampling. Default is 4000.}

\item{chunk_max}{Maximum chunk size for bootstrap sampling. Default is 2e6.}

\item{max_draws}{Maximum total number of bootstrap draws. Default is Inf.}

\item{patience_noaccept}{Number of consecutive chunks with no accepted samples before stopping. Default is 5.}
}
\value{
A list containing the point estimate, variance, number of kept samples, total draws, number of attempts, and status.
}
\description{
Computes a SAFE bootstrap estimate of the effect size and its variance for dependent samples.
}
