% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submerge.R
\name{submerge}
\alias{submerge}
\title{submerge}
\usage{
submerge(object1, object2, ..., mix = FALSE)
}
\arguments{
\item{object1}{object of class \code{orchard}.}

\item{object2}{object of class \code{orchard}.}

\item{...}{Other arguments passed to submerge.}

\item{mix}{If \code{TRUE}, it will add the number to the moderator name.}
}
\value{
Returns a data frame.
}
\description{
Merge two model results tables (orchard objects).
}
\author{
Shinichi Nakagawa - s.nakagawa@unsw.edu.au

Daniel Noble - daniel.noble@anu.edu.au
}
