% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf.R
\name{transf_ft}
\alias{transf_ft}
\title{Freeman-Tukey (Double Arcsine) Transformation for proportions}
\usage{
transf_ft(x, n)
}
\arguments{
\item{x}{A numeric vector containing proportions.}

\item{n}{A numeric value representing the sample size for proportion.}
}
\value{
Back transformed proportion.
}
\description{
Freeman-Tukey (Double Arcsine) Transformation for proportions
}
\keyword{internal}
