% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf.R
\name{transf_inv_ft}
\alias{transf_inv_ft}
\title{Inverse Freeman-Tukey (Double Arcsine) Transformation for proportions}
\usage{
transf_inv_ft(x, n)
}
\arguments{
\item{x}{A numeric vector containing effect sizes.}

\item{n}{A numeric value representing the sample size for proportion.}
}
\value{
Back transformed proportion.
}
\description{
Inverse Freeman-Tukey (Double Arcsine) Transformation for proportions
}
\keyword{internal}
