% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf.R
\name{transf_invlogit}
\alias{transf_invlogit}
\title{Inverse Logit Transformation}
\usage{
transf_invlogit(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The inverse logit of x, computed as exp(x) / (1 + exp(x)).
}
\description{
Inverse Logit Transformation
}
\keyword{internal}
