% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf.R
\name{transf_percent}
\alias{transf_percent}
\title{Percentage Transformation}
\usage{
transf_percent(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The percentage transformation, computed as exp(x) * 100.
}
\description{
Percentage Transformation
}
\keyword{internal}
