% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf.R
\name{transf_percentr}
\alias{transf_percentr}
\title{Percentage Relative Change Transformation}
\usage{
transf_percentr(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The percentage relative change transformation, computed as (exp(x) - 1) * 100.
}
\description{
Percentage Relative Change Transformation
}
\keyword{internal}
