% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf.R
\name{transf_tanh}
\alias{transf_tanh}
\title{Hyperbolic Tangent Transformation}
\usage{
transf_tanh(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The hyperbolic tangent of x.
}
\description{
Hyperbolic Tangent Transformation
}
\keyword{internal}
