% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf.R
\name{transform_data}
\alias{transform_data}
\title{Apply a Transformation to Vector}
\usage{
transform_data(
  x,
  n = NULL,
  transfm = c("none", "tanh", "invlogit", "percentr", "percent", "inv_ft")
)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{Used with "inv_ft". A numeric value representing the sample size for proportion.}

\item{transfm}{A character string specifying the transformation to apply. 
Options include:
- "none": No transformation (default)
- "tanh": Hyperbolic tangent transformation
- "invlogit": Inverse logit transformation
- "percentr": Percentage relative change transformation
- "percent": Percentage transformation
- "inv_ft": Inverse Freeman-Tukey (double arcsine) transformation for proportions (not yet working)}
}
\value{
A numeric vector with the applied transformation.
}
\description{
This function applies a specified transformation to a numeric vector.
}
\examples{
\dontrun{
transform_data(c(-1, 0, 1), transfm = "tanh")
transform_data(c(-1, 0, 1), transfm = "invlogit")
transform_data(c(-1, 0, 1), transfm = "none")
}
}
\keyword{internal}
