% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VCV_dARR}
\alias{VCV_dARR}
\title{VCV_dARR}
\format{
A sampling variance matrix :
}
\description{
Sampling error matrix used in Pottier et al. (2021) meta-analysis on sex-differences in acclimation
}
\references{
Pottier et al. 2022. Functional Ecology
}
