% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{weighted_var}
\alias{weighted_var}
\title{weighted_var}
\usage{
weighted_var(x, weights)
}
\arguments{
\item{x}{A vector of tau2s to be averaged}

\item{weights}{Weights, or sample sizes, used to average the variance}
}
\value{
Returns a vector with a single weighted variance
}
\description{
Calculate weighted variance
}
\author{
Shinichi Nakagawa - s.nakagawa@unsw.edu.au

Daniel Noble - daniel.noble@anu.edu.au
}
