% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchard_plot.R
\name{Zr_to_r}
\alias{Zr_to_r}
\title{Zr_to_r}
\usage{
Zr_to_r(df)
}
\arguments{
\item{df}{data frame of results of class 'orchard'}
}
\value{
A data frame containing all the model results including mean effect size estimate, confidence and prediction intervals with estimates converted back to r
}
\description{
Converts Zr back to r (Pearson's correlation coefficient)
}
\author{
Shinichi Nakagawa - s.nakagawa@unsw.edu.au

Daniel Noble - daniel.noble@anu.edu.au
}
