% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{bubble_plot}
\alias{bubble_plot}
\title{bubble_plot}
\usage{
bubble_plot(
  object,
  mod,
  group,
  data,
  xlab = "Moderator",
  ylab = "Effect size",
  N = "none",
  alpha = 0.5,
  cb = TRUE,
  k = TRUE,
  g = TRUE,
  est.lwd = 1,
  ci.lwd = 0.5,
  pi.lwd = 0.5,
  est.col = "black",
  ci.col = "black",
  pi.col = "black",
  legend.pos = c("top.right", "top.left", "bottom.right", "bottom.left", "top.out",
    "bottom.out"),
  condition.nrow = 2,
  weights = "prop",
  by = NULL,
  at = NULL
)
}
\arguments{
\item{object}{model object of class 'rma.mv', 'rma' or 'orchard' table of model results}

\item{mod}{the name of a continious moderator.}

\item{group}{The grouping variable that one wishes to plot beside total effect sizes, k. This could be study, species or whatever other grouping variable one wishes to present sample sizes. Not needed of a orchard_plot is provided with a mod_results object of class 'orchard'.}

\item{data}{The data frame used to fit the rma.mv model object. Not needed of a orchard_plot is provided with a mod_results object of class 'orchard'.}

\item{xlab}{Effect size measure label.}

\item{N}{The vector of sample size which an effect size is based on. If default, we use precision (the inverse of sampling standard error)}

\item{alpha}{The level of transparency for pieces of fruit (effect size)}

\item{cb}{If TRUE, it uses 20 colour blind friendly colors}

\item{k}{If TRUE, it displays k (number of effect sizes) on the plot}

\item{g}{If TRUE, it displays g (number of grouping levels for each level of the moderator) on the plot}

\item{weights}{Used when one wants marginalised means. How to marginalize categorical variables. The default is weights = "prop", which wights means for moderator levels based on their proportional representation in the data. For example, if "sex" is a moderator, and males have a larger sample size than females, then this will produce a weighted average, where males are weighted more towards the mean than females. This may not always be ideal. IN the case if sex, for example, males and females are roughly equally prevalent in a population. As such, you can give the moderator levels equal weight using weights = "equal".}

\item{by}{Used when one wants marginalised (mearginal) means. The 'condition' variable that one wishes to have the mean for the moderator vary.}

\item{at}{Used when one wants marginalised means. The 'condition' that one wishes to calculate the means at, but is not presented in output}
}
\value{
Orchard plot
}
\description{
Using a metafor model object of class rma or rma.mv or a results table of class orchard, it creates a an bubble plot from slope estimates or panels of bubble plot in cases when a model includes interaction terms.
}
\examples{
\dontrun{
data(eklof)

}
}
\author{
Shinichi Nakagawa - s.nakagawa@unsw.edu.au

Daniel Noble - daniel.noble@anu.edu.au
}
