% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{bubble_plot}
\alias{bubble_plot}
\title{bubble_plot}
\usage{
bubble_plot(
  object,
  mod,
  group = NULL,
  xlab = "Moderator",
  ylab = "Effect size",
  N = "none",
  alpha = 0.5,
  cb = TRUE,
  k = TRUE,
  g = FALSE,
  transfm = c("none", "tanh", "invlogit", "percent", "percentr"),
  n_transfm = NULL,
  est.lwd = 1,
  ci.lwd = 0.5,
  pi.lwd = 0.5,
  est.col = "black",
  ci.col = "black",
  pi.col = "black",
  legend.pos = c("top.left", "top.right", "bottom.right", "bottom.left", "top.out",
    "bottom.out", "none"),
  k.pos = c("top.right", "top.left", "bottom.right", "bottom.left", "none"),
  condition.nrow = 2,
  weights = "prop",
  by = NULL,
  at = NULL,
  condition.levels = NULL
)
}
\arguments{
\item{object}{Model object of class \code{rma}, \code{rma.mv}, or \code{orchard} table of model results}

\item{mod}{The name of a continuous moderator, to be plotted on the x-axis of the bubble plot.}

\item{group}{The grouping variable that one wishes to plot beside total effect sizes, k. This could be study, species, or any grouping variable one wishes to present sample sizes for. Not needed if an \code{orchard_plot} is provided with a \code{mod_results} object of class \code{orchard}.}

\item{xlab}{Moderator label.}

\item{ylab}{Effect size measure label.}

\item{N}{The vector of sample size which an effect size is based on. Defaults to precision (the inverse of sampling standard error).}

\item{alpha}{The level of transparency for pieces of fruit (effect size).}

\item{cb}{If \code{TRUE}, it uses a colourblind-friendly palette of 20 colours (do not make this \code{TRUE}, when colour = \code{TRUE}).}

\item{k}{If \code{TRUE}, it displays k (number of effect sizes) on the plot.}

\item{g}{If \code{TRUE}, it displays g (number of grouping levels for each level of the moderator) on the plot.}

\item{transfm}{If set to \code{"tanh"}, a tanh transformation will be applied to effect sizes, converting Zr to a correlation or pulling in extreme values for other effect sizes (lnRR, lnCVR, SMD).  \code{"invlogit"} can be used to convert lnRR to the inverse logit scale. \code{"percentr"} can convert to the percentage change scale when using response ratios and \code{"percent"} can convert to the percentage change scale of an log transformed effect size. Defaults to \code{"none"}.}

\item{n_transfm}{The vector of sample sizes for each effect size estimate. This is used when \code{transfm = "inv_ft"}. Defaults to NULL.}

\item{est.lwd}{Size of the point estimate.}

\item{ci.lwd}{Size of the confidence interval.}

\item{pi.lwd}{Size of the prediction interval.}

\item{est.col}{Colour of the point estimate.}

\item{ci.col}{Colour of the confidence interval.}

\item{pi.col}{Colour of the prediction interval.}

\item{legend.pos}{Where to place the legend, or not to include a legend ("none").}

\item{k.pos}{The position of effect size number, k.}

\item{condition.nrow}{Number of rows to plot condition variable.}

\item{weights}{How to marginalize categorical variables; used when one wants marginalised means. The default is \code{weights = "prop"}, which weights means for moderator levels based on their proportional representation in the data. For example, if \code{"sex"} is a moderator, and males have a larger sample size than females, then this will produce a weighted average, where males are weighted more towards the mean than females. This may not always be ideal when, for example, males and females are typically roughly equally prevalent in a population. In cases such as these, you can give the moderator levels equal weight using \code{weights = "equal"}.}

\item{by}{Character vector indicating the name that predictions should be conditioned on for the levels of the moderator.}

\item{at}{List of levels one wishes to predict at for the corresponding variables in \code{by}. Used when one wants marginalised means. This argument can also be used to suppress levels of the moderator when argument \code{subset = TRUE}. Provide a list as follows: \code{list(mod = c("level1", "level2"))}.}

\item{condition.levels}{Order of the levels of the condition variable in the order to plot. Defaults to NULL.}
}
\value{
Bubble plot
}
\description{
Using a \pkg{metafor} model object of class \code{rma} or \code{rma.mv}, or a results table of class \code{orchard}, the \code{bubble_plot} function creates a bubble plot from slope estimates. In cases when a model includes interaction terms, this function creates panels of bubble plots.
}
\examples{
\dontrun{
data(lim)
lim[, "year"] <- as.numeric(lim$year)
lim$vi <- 1 / (lim$N - 3)
model <- metafor::rma.mv(
  yi = yi, V = vi, mods = ~ Environment * year,
  random = list(~ 1 | Article, ~ 1 | Datapoint), data = na.omit(lim)
)
test <- orchaRd::mod_results(model, mod = "year", group = "Article", data = lim, weights = "prop", by = "Environment")
orchaRd::bubble_plot(test, mod = "year", group = "Article", legend.pos = "top.left")
# Or just using model directly
orchaRd::bubble_plot(model, mod = "year", legend.pos = "top.left", group = "Article", weights = "prop", by = "Environment")
}
}
\author{
Shinichi Nakagawa - s.nakagawa@unsw.edu.au

Daniel Noble - daniel.noble@anu.edu.au
}
