% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caterpillars.R
\name{caterpillars}
\alias{caterpillars}
\title{caterpillars}
\usage{
caterpillars(
  object,
  mod = "Int",
  xlab,
  overall = TRUE,
  transfm = c("none", "tanh")
)
}
\arguments{
\item{object}{Model object of class 'rma.mv', 'rma' or 'orchard' table of model results}

\item{mod}{The name of a moderator. Otherwise, "Int" for intercept only model.}

\item{xlab}{The effect size measure label.}

\item{overall}{Logical indicating whether to relabel "Intrcpt" (the default label from rma or rma.mv intercept only models or meta-analyses) to "Overall".}

\item{transfm}{If set to "tanh", a tanh transformation will be applied to effect sizes, converting Zr will to a correlation or pulling in extreme, values for other effect sizes (lnRR, lnCVR, SMD). If "none" is chosen then it will default.}
}
\value{
Caterpillars plot
}
\description{
Using a metafor model object of class rma or rma.mv or a results table of class orchard, it creates a an caterpillars plot from mean effect size estimates for all levels of a given categorical moderator, their corresponding confidence intervals and prediction intervals
}
\examples{
\dontrun{
data(eklof)
eklof<-metafor::escalc(measure="ROM", n1i=N_control, sd1i=SD_control,
m1i=mean_control, n2i=N_treatment, sd2i=SD_treatment, m2i=mean_treatment,
data=eklof)
# Add the unit level predictor
eklof$Datapoint<-as.factor(seq(1, dim(eklof)[1], 1))
# fit a MLMR - accouting for some non-independence
eklof_MR<-metafor::rma.mv(yi=yi, V=vi, mods=~ Grazer.type-1, random=list(~1|ExptID,
~1|Datapoint), data=eklof)
results <- mod_results(eklof_MR, mod = "Grazer.type")
caterpillars(results, mod = Grazer.type, xlab = "log(Response ratio) (lnRR)")
# or
caterpillars(eklof_MR, mod = Grazer.type, xlab = "log(Response ratio) (lnRR)")

# Example 2
data(lim)
lim$vi<- 1/(lim$N - 3)
lim_MR<-metafor::rma.mv(yi=yi, V=vi, mods=~Phylum-1, random=list(~1|Article,
~1|Datapoint), data=lim)
caterpillars(lim_MR, mod = "Phylum", xlab = "Correlaiton coefficent", transfm = "tanh", N = lim$N)
}
}
\author{
Shinichi Nakagawa - s.nakagawa@unsw.edu.au

Daniel Noble - daniel.noble@anu.edu.au
}
