% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leave_one_out.R
\name{.create_tmp_phylo_matrix}
\alias{.create_tmp_phylo_matrix}
\title{Create Temporary Phylogenetic Matrix For Leave-One-Out}
\usage{
.create_tmp_phylo_matrix(data, phylo_args)
}
\arguments{
\item{data}{A data frame containing the variables specified in phylo_args.}

\item{phylo_args}{A list of arguments for the phylogenetic matrix calculation, including:
\itemize{
  \item tree: A phylogenetic tree object
  \item species_colname: Name of the column in the data that is linked to the phylo matrix (species names)
}}
}
\value{
A phylogenetic matrix for use in meta-analytic models.
}
\description{
Create Temporary Phylogenetic Matrix For Leave-One-Out
}
\author{
Facundo Decunta -
}
\keyword{internal}
