% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{.get_kg_labels}
\alias{.get_kg_labels}
\title{Compute k and g labels for each condition.}
\usage{
.get_kg_labels(data_trim)
}
\description{
Compute k and g labels for each condition.
}
\keyword{internal}
