% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchard_plot.R
\name{.orcd_point_estimates}
\alias{.orcd_point_estimates}
\title{Add point estimates for orchard plot}
\usage{
.orcd_point_estimates(
  mod_table,
  estimates_fill,
  points_shapes,
  lines_position,
  trunk.size,
  shapes_values
)
}
\description{
Add point estimates for orchard plot
}
\keyword{internal}
