% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leave_one_out.R
\name{.run_leave1out}
\alias{.run_leave1out}
\title{Fit Multiple Meta-Analytic Models For Leave-One-Out Analysis}
\usage{
.run_leave1out(
  model,
  group,
  vcalc_args = NULL,
  robust_args = NULL,
  phylo_args = NULL
)
}
\arguments{
\item{model}{A fitted model object containing a \code{data} element, which must be 
a data frame with all model variables.}

\item{group}{A character string specifying the column in \code{model$data} that 
defines the groups to be omitted one at a time.}

\item{vcalc_args}{Optional list of arguments for the variance-covariance calculation using 
metafor's vcalc function. Must include 'vi', 'cluster', 'obs', and 'rho' elements.}

\item{robust_args}{Optional list of arguments for robust variance estimation using
metafor's robust function. Must include a 'cluster' element.}

\item{phylo_args}{Optional list of arguments for phylogenetic matrix calculation using
ape's vcv function. Must include 'tree' and 'species_colname' elements. 'tree' is a phylogenetic
tree object and 'species_colname' is the name of the column in the model data that is linked to the phylo matrix.}
}
\value{
A named list of models, each fitted after omitting one group. Names correspond 
  to the omitted group IDs.
}
\description{
Iteratively refits a meta-analytic model, leaving out one level of a specified 
grouping variable in each iteration.
}
\details{
The function removes each unique group from \code{model$data} one at a time, 
  refitting the model using \code{update()}. If an update fails, a warning is issued, 
  and \code{NULL} is returned for that group.
}
\author{
Facundo Decunta - fdecunta@agro.uba.ar
}
\keyword{internal}
