% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{english}
\alias{english}
\title{english}
\format{A data frame :
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{}
}
\description{
Dataset from English and Uller 2012  and Senior et al. 2017 meta-analyses.
}
