% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{firstup}
\alias{firstup}
\title{firstup}
\usage{
firstup(x)
}
\arguments{
\item{x}{a character string}
}
\value{
Returns a character string with all combinations of the moderator level names with upper case first letters
}
\description{
Uppercase moderator names
}
