% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{get_data_raw}
\alias{get_data_raw}
\title{get_data_raw}
\usage{
get_data_raw(model, mod, group, data, at = NULL)
}
\arguments{
\item{model}{rma.mv object}

\item{mod}{the moderator variable}

\item{group}{The grouping variable that one wishes to plot beside total effect sizes, k. This could be study, species or whatever other grouping variable one wishes to present sample sizes.}

\item{data}{The data frame used to fit the rma.mv model object}

\item{at}{If at is moderator levels then at will be used to subset levels.}
}
\value{
Returns a data frame
}
\description{
Collects and builds the data used to fit the rma.mv or rma model in metafor
}
\examples{
\dontrun{
data(fish)
warm_dat <- fish
model <- metafor::rma.mv(yi = lnrr, V = lnrr_vi, random = list(~1 | group_ID, ~1 | es_ID), mods = ~ experimental_design + trait.type + deg_dif + treat_end_days, method = "REML", test = "t", data = warm_dat, control=list(optimizer="optim", optmethod="Nelder-Mead"))
 test <- get_data_raw(model, mod = "trait.type", group = "group_ID", data = warm_dat, at = list(trait.type = c("physiology", "morphology")))
 test2 <- get_data_raw(model, mod = "1", group = "group_ID", data = warm_dat)

 data(english)
 # We need to calculate the effect sizes, in this case d
 english <- escalc(measure = "SMD", n1i = NStartControl, sd1i = SD_C, m1i = MeanC, n2i = NStartExpt, sd2i = SD_E, m2i = MeanE, var.names=c("SMD","vSMD"), data = english)
 model <- rma.mv(yi = SMD, V = vSMD, random = list( ~ 1 | StudyNo, ~ 1 | EffectID), data = english)
 test3 <-  get_data_raw(model, mod = "1", group = "StudyNo", data = english)}
}
\author{
Shinichi Nakagawa - s.nakagawa@unsw.edu.au

Daniel Noble - daniel.noble@anu.edu.au
}
