% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i2_ml.R
\name{i2_ml}
\alias{i2_ml}
\title{i2_ml}
\usage{
i2_ml(model, method = c("ns", "wv"))
}
\arguments{
\item{model}{Model object of class 'rma.mv', 'rma'}

\item{method}{Method used to calculate I2. Two options exist, Nakagawa & Santos ('ns') or Wolfgang Viechtbauer's method ("wv").}
}
\value{
A data frame containing all the model results including mean effect size estimate, confidence and prediction intervals with estimates converted back to r
}
\description{
I2 (I-squared) for mulilevel meta-analytic models, based on Nakagawa & Santos (2012). Under multilevel models, we can have a multiple I2 (see also Senior et al. 2016). Alternatively, the method proposed by Wolfgang Viechtbauer (http://www.metafor-project.org/doku.php/tips:i2_multilevel_multivariate?s[]=multilevel) can also be used.
}
\examples{
\dontrun{
data(english)
english <- escalc(measure = "SMD", n1i = NStartControl, sd1i = SD_C, m1i = MeanC, n2i = NStartExpt, sd2i = SD_E, m2i = MeanE, var.names=c("SMD","vSMD"),data = english)
english_MA <- rma.mv(yi = SMD, V = vSMD, random = list( ~ 1 | StudyNo, ~ 1 | EffectID), data = english)
I2 <- i2_ml(english_MA)
}
}
\references{
Senior, A. M., Grueber, C. E., Kamiya, T., Lagisz, M., O’Dwyer, K., Santos, E. S. A. & Nakagawa S. 2016. Heterogeneity in ecological and evolutionary meta-analyses: its magnitudes and implications. *Ecology* 97(12): 3293-3299.

Nakagawa, S, and Santos, E.S.A. 2012. Methodological issues and advances in biological meta-analysis. *Evolutionary Ecology* 26(5): 1253-1274.
}
\author{
Shinichi Nakagawa - s.nakagawa@unsw.edu.au

Daniel Noble - daniel.noble@anu.edu.au
}
