% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leave_one_out.R
\name{leave_one_out}
\alias{leave_one_out}
\title{Leave-One-Out Analysis for Meta-Analytic Models}
\usage{
leave_one_out(model, group, vcalc_args = NULL, robust_args = NULL)
}
\arguments{
\item{model}{A meta-analytic model fitted using **metafor**.}

\item{group}{A factor or categorical variable specifying the leave-one-out groups.}
}
\value{
Same as `mod_results()`, but with the estimates from each model ran
  in the leave-one-out analysis, and the effect sizes from each model.
}
\description{
Performs a leave-one-out (LOO) analysis on a meta-analytic model from
the **metafor** package by iteratively removing each level of a grouping
variable and refitting the model.
}
\examples{
\dontrun{
res <- metafor::rma.mv(lnrr, lnrr_vi, random = ~ 1 | paper_ID, data = fish)
loo_results <- leave_one_out(res, group = "paper_ID")
}

}
\author{
Facundo Decunta - fdecunta@agro.uba.ar
}
