% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leave_one_out.R
\name{leave_one_out}
\alias{leave_one_out}
\title{Leave-One-Out Analysis for Meta-Analytic Models}
\usage{
leave_one_out(
  model,
  group,
  vcalc_args = NULL,
  robust_args = NULL,
  phylo_args = NULL
)
}
\arguments{
\item{model}{A meta-analytic model fitted using **metafor** package functions such as rma.mv().}

\item{group}{A character string specifying the column name in model$data that contains the 
grouping variable for which levels will be iteratively removed.}

\item{vcalc_args}{Optional list of arguments for variance-covariance calculation using 
metafor's vcalc function. Must include:
\itemize{
  \item \code{vi}: Name of the variance column
  \item \code{cluster}: Name of the clustering variable column
  \item \code{obs}: Name of the observation ID column
  \item \code{rho}: Correlation coefficient between effect sizes
}}

\item{robust_args}{Optional list of arguments for robust variance estimation using
metafor's robust function. Must include:
\itemize{
  \item \code{cluster}: Name of the clustering variable column
  \item \code{clubSandwich}: Logical indicating whether to use clubSandwich method (optional)
}}

\item{phylo_args}{Optional list of arguments for phylogenetic matrix calculation using
ape's vcv function. Must include:
\itemize{
  \item \code{tree}: A phylogenetic tree object of class "phylo"
  \item \code{species_colname}: Name of the column in model data linked to the tree tips
}}
}
\value{
An object of class "orchard" containing:
  \itemize{
    \item \code{mod_table}: A data frame with model estimates from each leave-one-out iteration,
          with an additional column indicating which group was omitted.
    \item \code{data}: A data frame with effect sizes from each iteration, with an additional
          column indicating which group was omitted.
    \item \code{orig_mod_results}: The results from the original model without any omissions,
          as returned by mod_results().
  }
}
\description{
Performs a leave-one-out analysis on a meta-analytic model from
the **metafor** package by iteratively removing each level of a grouping
variable and refitting the model.
}
\examples{
\dontrun{
res <- metafor::rma.mv(lnrr, lnrr_vi, random = ~ 1 | paper_ID, data = fish)
loo_results <- leave_one_out(res, group = "paper_ID")

# With variance-covariance calculation
loo_results <- leave_one_out(res, group = "paper_ID", 
                             vcalc_args = list(vi = "lnrr_vi", 
                                              cluster = "paper_ID", 
                                              obs = "es_ID", 
                                              rho = 0.5))
                                              
# With robust variance estimation
loo_results <- leave_one_out(res, group = "paper_ID", 
                             robust_args = list(cluster = "paper_ID"))
}

}
\author{
Facundo Decunta - fdecunta@agro.uba.ar
}
