% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lim}
\alias{lim}
\title{lim}
\format{
A data frame :
\describe{
  \item{Article}{Study or research paper that data were collected}
  \item{Datapoint}{Observation level ID that identified each unique datapoint}
\item{N}{Sample size used to estaimted correlation}
\item{Phylum}{Phylum of the species}
\item{yi}{Effect size correlation coefficient between maternal and offspring size within-species}
\item{Author}{Authors of article}
  ...
}
}
\description{
Lim et al. (2014) meta-analysed the strength of correlation between maternal and offspring size within-species, across a very wide range of taxa. They found, that typically, there is a moderate positive correlation between maternal size and offspring size within species (i.e. larger mothers have larger offspring). Below we describe the variables use in our examples and or analyses.
}
\references{
Lim J.N., Senior A.M., Nakagawa S. 2014. Heterogeneity in individual quality and reproductive trade-offs within species. Evolution. 68(8):2306-18. doi: 10.1111/evo.12446
}
