% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lim}
\alias{lim}
\title{lim}
\format{A data frame :
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\description{
Dataset from Lim et al. 2014 meta-analyses.
}
