% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_stats.R
\name{magnitude_effects}
\alias{magnitude_effects}
\title{Magnitude Effects using SAFE Bootstrap}
\usage{
magnitude_effects(
  x1bar,
  x2bar,
  sd1,
  sd2,
  n1,
  n2,
  min_kept = 2000,
  chunk_init = 4000,
  chunk_max = 2e+06,
  max_draws = Inf,
  patience_noaccept = 5,
  paired = FALSE,
  r = NULL,
  seed = 565
)
}
\arguments{
\item{x1bar}{Mean of group 1.}

\item{x2bar}{Mean of group 2.}

\item{sd1}{Standard deviation of group 1.}

\item{sd2}{Standard deviation of group 2.}

\item{n1}{Sample size of group 1.}

\item{n2}{Sample size of group 2.}

\item{min_kept}{Minimum number of accepted bootstrap samples to keep. Default is 2000.}

\item{chunk_init}{Initial chunk size for bootstrap sampling. Default is 4000.}

\item{chunk_max}{Maximum chunk size for bootstrap sampling. Default is 2e6.}

\item{max_draws}{Maximum total number of bootstrap draws. Default is Inf.}

\item{patience_noaccept}{Number of consecutive chunks with no accepted samples before stopping. Default is 5.}

\item{paired}{Logical, whether the samples are paired. Default is FALSE.}

\item{r}{Optional correlation between groups for dependent samples. Required if `paired` is TRUE.}

\item{seed}{Random seed for reproducibility. Default is 565.}
}
\value{
A list containing the point estimate, variance, number of kept samples, total draws, number of attempts, and status.
}
\description{
Computes the magnitude effect size and its variance using the SAFE bootstrap method for either independent or dependent samples.
}
\examples{
\dontrun{
# Independent samples
magnitude_effects(x1bar = 10, x2bar = 7, sd1 = 2, sd2 = 3, n1 = 30, n2 = 30)
# Dependent samples with correlation
magnitude_effects(x1bar = 10, x2bar = 7, sd1 = 2, sd2 = 3, n1 = 30, n2 = 30, paired = TRUE, r = 0.5)
# Example across rows of data frame
data = data.frame(xbar1 = rep(5, 5), xbar2 = rep(20, 5), sd1= rep(5, 5), sd2= rep(6, 5), n1= rep(30, 5), n2= rep(30, 5))
res <- data \%>\%
  mutate(
    mag = purrr::pmap(
      list(xbar1, xbar2, sd1, sd2, n1, n2),
      ~ magnitude_effects(
          x1bar = ..1, x2bar = ..2,
          sd1 = ..3, sd2 = ..4,
          n1 = ..5, n2 = ..6,
          paired = FALSE))) \%>\%
  unnest_wider(mag)

# Alternatively, you can use `rowwise()` 
res <- data \%>\%
  rowwise() \%>\%
  mutate(out = list(magnitude_effects(xbar1, xbar2, sd1, sd2, n1, n2,
                                   min_kept = 200, chunk_init = 400))) \%>\%
  unnest_wider(out)

# Paired, use add correlation to dataframe
data = data  \%>\% mutate(r = 0.4)

res2 <- data \%>\%
  mutate(
    mag = purrr::pmap(
      list(xbar1, xbar2, sd1, sd2, n1, n2, r),
      ~ magnitude_effects(
          x1bar = ..1, x2bar = ..2,
          sd1 = ..3, sd2 = ..4,
          n1 = ..5, n2 = ..6, r = ..7,
          paired = TRUE))) \%>\%
  unnest_wider(mag)

}
}
