% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{marginal_means}
\alias{marginal_means}
\title{marginal_means}
\usage{
marginal_means(
  model,
  data,
  mod = "1",
  weights = "prop",
  by = NULL,
  at = NULL,
  ...
)
}
\arguments{
\item{model}{rma.mv object}

\item{data}{data frame used to fit rma.mv model}

\item{mod}{moderator variable of interest that one wants marginal means for.}

\item{weights}{how to marginalize categorical variables. The default is weights = "prop" - more to add here (discuss with Dan)}
}
\description{
Function to to get marginal means from meta-regression models with single or multiple moderator variables that are both continuous or categorical.
}
\examples{
\dontrun{
data(fish)
warm_dat <- fish
model <- metafor::rma.mv(yi = lnrr, V = lnrr_vi, random = list(~1 | group_ID, ~1 | es_ID), mods = ~ experimental_design + trait.type + deg_dif + treat_end_days, method = "REML", test = "t", data = warm_dat, control=list(optimizer="optim", optmethod="Nelder-Mead"))
  overall <- marginal_means(model, data = warm_dat)
across_trait <- marginal_means(model, data = warm_dat, pred = "trait.type")
across_trait_by_degree_diff <- marginal_means(model, data = warm_dat, mod = "trait.type", at = list(deg_dif = c(5, 10, 15)), by = "deg_dif")
across_trait_by_degree_diff_at_treat_end_days10 <- marginal_means(model, data = warm_dat, mod = "trait.type", at = list(deg_dif = c(5, 10, 15), treat_end_days = 10), by = "deg_dif")
across_trait_by_degree_diff_at_treat_end_days10And50 <- marginal_means(model, data = warm_dat, mod = "trait.type", at = list(deg_dif = c(5, 10, 15), treat_end_days = c(10, 50)), by = "deg_dif")
across_trait_by_treat_end_days10And50 <- marginal_means(model, data = warm_dat, mod = "trait.type", at = list(deg_dif = c(5, 10, 15), treat_end_days = c(10, 50)), by = "treat_end_days")
across_trait_by_treat_end_days10And50_ordinaryMM <- marginal_means(model, data = warm_dat, mod = "trait.type", at = list(deg_dif = c(5, 10, 15), treat_end_days = c(10, 50)), by = "treat_end_days", weights = "prop")
}
}
\author{
Shinichi Nakagawa - s.nakagawa@unsw.edu.au

Daniel Noble - daniel.noble@anu.edu.au
}
