% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvh1_ml.R
\name{ml_cvh1}
\alias{ml_cvh1}
\title{ml_cvh1}
\usage{
ml_cvh1(model)
}
\arguments{
\item{model}{Model object of class \code{rma.mv} or \code{rma}.}
}
\description{
Calculated CVH1 for mulilevel meta-analytic models
}
