% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvh2_ml.R
\name{ml_cvh2}
\alias{ml_cvh2}
\title{ml_cvh2}
\usage{
ml_cvh2(model)
}
\arguments{
\item{model}{Model object of class \code{rma.mv} or \code{rma}.}
}
\description{
Calculated CVH2 for mulilevel meta-analytic models
}
