% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{mod_is_valid}
\alias{mod_is_valid}
\title{mod_is_valid}
\usage{
mod_is_valid(model, mod)
}
\arguments{
\item{model}{A meta-analytic model from the \code{metafor} package.}

\item{mod}{A string specifying the moderator to check. Must be `"1"`
or one of the moderators included in `model$formula.mods`.}
}
\value{
A logical value: `TRUE` if `mod` is valid, `FALSE` otherwise.
}
\description{
Checks whether the specified `mod` argument is valid.
It must be either `"1"` (indicating an intercept-only model) or
a moderator included in the `model` object.
}
\details{
Check if the specified moderator is valid.
}
\keyword{internal}
