% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{mod_results}
\alias{mod_results}
\title{mod_results}
\usage{
mod_results(model, mod = "")
}
\arguments{
\item{model}{rma.mv object}

\item{mod}{the name of a moderator}
}
\value{
A data frame containing all the model results including mean effect size estimate, confidence and prediction intervals

Data table (tibble)
}
\description{
Using a metafor model object of class rma or rma.mv it creates a table of model results containing the mean effect size estimates for all levels of a given categorical moderator, their corresponding confidence intervals and prediction intervals
}
\examples{
\dontrun{data(eklof) 
eklof<-metafor::escalc(measure="ROM", n1i=N_control, sd1i=SD_control,
m1i=mean_control, n2i=N_treatment, sd2i=SD_treatment, m2i=mean_treatment,
data=eklof)
# Add the unit level predictor
eklof$Datapoint<-as.factor(seq(1, dim(eklof)[1], 1))
# fit a MLMR - accouting for some non-independence
eklof_MR<-metafor::rma.mv(yi=yi, V=vi, mods=~ Grazer.type-1, random=list(~1|ExptID,
~1|Datapoint), data=eklof)
results <- mod_results(eklof_MR, mod = "Grazer.type")
}
}
