% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchard_plot.R
\name{orchard_plot}
\alias{orchard_plot}
\title{orchard_plot}
\usage{
orchard_plot(object, mod, es_type = c("d", "Zr", "lnRR", "lnCVR"),
  alpha = 0.8, N = "none")
}
\arguments{
\item{object}{object of class 'rma.mv', 'rma' or 'orchard '}

\item{mod}{the name of a moderator}

\item{es_type}{the type of effect size used in the model, z-transformed Persons correlation coefficient (Zr) or standardised mean difference (i.e., Hedges/Cohen's d, g) (d)}

\item{N}{The vector of sample size}
}
\value{
Orchard plot
}
\description{
Using a metafor model object of class rma or rma.mv or a results table of class orchard, it creates a an orchard plot from mean effect size estimates for all levels of a given categorical moderator, their corresponding confidence intervals and prediction intervals
}
\examples{
\dontrun{
data(eklof) 
eklof<-metafor::escalc(measure="ROM", n1i=N_control, sd1i=SD_control,
m1i=mean_control, n2i=N_treatment, sd2i=SD_treatment, m2i=mean_treatment,
data=eklof)
# Add the unit level predictor
eklof$Datapoint<-as.factor(seq(1, dim(eklof)[1], 1))
# fit a MLMR - accouting for some non-independence
eklof_MR<-metafor::rma.mv(yi=yi, V=vi, mods=~ Grazer.type-1, random=list(~1|ExptID,
~1|Datapoint), data=eklof)
results <- mod_results(eklof_MR, mod = "Grazer.type")
orchard_plot(results, data = eklof, mod = Grazer.type, es_type = "Zr")
# or
orchard_plot(eklof_MR, data = eklof, mod = Grazer.type, es_type = "Zr")

# Example 2
data(lim)
lim$vi<-(1/sqrt(lim$N - 3))^2
lim_MR<-metafor::rma.mv(yi=yi, V=vi, mods=~Phylum-1, random=list(~1|Article,
~1|Datapoint), data=lim)
orchard_plot(lim_MR, data = lim, mod = "Phylum", es_type = "Zr")
}
}
