% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{pred_interval_esmeans}
\alias{pred_interval_esmeans}
\title{pred_interval_esmeans}
\usage{
pred_interval_esmeans(model, esmeans)
}
\arguments{
\item{model}{rma.mv object}

\item{esmeans}{result from emmeans::emmeans object}
}
\description{
Function to get prediction intervals (crediblity intervals) from esmeans objects (metafor)
}
\author{
Shinichi Nakagawa - s.nakagawa@unsw.edu.au

Daniel Noble - daniel.noble@anu.edu.au
}
