% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf.R
\name{transform_data}
\alias{transform_data}
\title{Apply a Transformation to Data}
\usage{
transform_data(
  x,
  transfm = c("none", "tanh", "invlogit", "percentr", "percent")
)
}
\arguments{
\item{x}{A numeric vector.}

\item{transfm}{A character string specifying the transformation to apply. 
Options include:
- "none": No transformation (default)
- "tanh": Hyperbolic tangent transformation
- "invlogit": Inverse logit transformation
- "percentr": Percentage relative change transformation
- "percent": Percentage transformation}
}
\value{
A numeric vector with the applied transformation.
}
\description{
This function applies a specified transformation to a numeric vector.
}
\examples{
transform_data(c(-1, 0, 1), "tanh")
transform_data(c(-1, 0, 1), "invlogit")
}
\keyword{internal}
